+++ CHANGELOG.md
@@ -51,7 +51,7 @@
   * `-common.storage.gcs.max-retries`
   * `-ruler-storage.gcs.max-retries`
 * [ENHANCEMENT] Usage-tracker: Improve first snapshot loading & rehash speed. #13284
-* [ENHANCEMENT] Ruler: Implemented `OperatorControllableErrorClassifier` for rule evaluation that allows the differentiation between reason `operator` and `user`-controllable rule evaluation errors. #13313
+* [ENHANCEMENT] Ruler: Rule evaluation failure metric `prometheus_rule_evaluation_failures_total` now include a `reason` label with values `operator` or `user` to distinguish between operator-controllable errors (storage failures, 5xx errors, rate limiting) and user-controllable errors (bad queries, validation errors, 4xx errors). #13313 
 * [BUGFIX] Compactor: Fix potential concurrent map writes. #13053
 * [BUGFIX] Query-frontend: Fix issue where queries sometimes fail with `failed to receive query result stream message: rpc error: code = Canceled desc = context canceled` if remote execution is enabled. #13084
 * [BUGFIX] Query-frontend: Fix issue where query stats, such as series read, did not include the parameters to the `histogram_quantile` and `histogram_fraction` functions if remote execution was enabled. #13084